% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose_text.R
\name{rose_text}
\alias{rose_text}
\title{Add Text to Rose map}
\usage{
rose_text(mapping = NULL, data = NULL, center = c(0, 0),
  outer.r = 1.05, n = 1000, round = 360, angle_string = NULL,
  angle_id = NULL, coord_string = NULL, coord_id = NULL,
  disappear_string = NULL, disappear_id = NULL, translate.1 = NULL,
  trans_h.2 = NULL, trans_v.3 = NULL, trans_al.4 = NULL,
  rotate.5 = 0, zoom.6 = NULL, mirror.7 = NULL, mirror_h.8 = NULL,
  mirror_v.9 = NULL, position = "identity", ..., parse = FALSE,
  check_overlap = FALSE, na.rm = FALSE, show.legend = NA,
  angle = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data, must be a dataframe}

\item{center}{center coordinate, must be two numbers}

\item{outer.r}{the radius of text circle, default is 1.05}

\item{n}{the number of points, which will build a circle to help locate the text}

\item{round}{the round of map, which should be 0-360, default is 360.}

\item{angle_string}{change angle by string, which should be s1;s2:angle}

\item{angle_id}{change angle by id, which can be 1:2;3:angle}

\item{coord_string}{change coordinate by string, which can be s1;s2:0.1;0.2}

\item{coord_id}{change coordinate by id, which can be 1:2;3:0.1;0.2}

\item{disappear_string}{disappear by string, which can be s1;s2}

\item{disappear_id}{disappear by id, which can be 1:2;3}

\item{translate.1}{Operation to whole text, which is should be two numbers for x and y values}

\item{trans_h.2}{Operation to whole text, which is should be one number}

\item{trans_v.3}{Operation to whole text, which is should be one number}

\item{trans_al.4}{Operation to whole text, which is should be two numbers for slope and intercept values}

\item{rotate.5}{Operation to whole text, which is should be one number for angle}

\item{zoom.6}{Operation to whole text, which is should be one number for ratio}

\item{mirror.7}{Operation to whole text, which is should be two numbers for slope and intercept values}

\item{mirror_h.8}{Operation to whole text, which is should be one number}

\item{mirror_v.9}{Operation to whole text, which is should be one number}

\item{position}{position}

\item{...}{passed to parameters}

\item{parse}{default is FALSE}

\item{check_overlap}{default is FALSE}

\item{na.rm}{default is FALSE}

\item{show.legend}{default is FALSE}

\item{angle}{default is NA}

\item{inherit.aes}{default is TRUE}
}
\value{
text
}
\description{
Add Text to Rose map
}
\examples{
library(ggrose)
set.seed(2020)
df=data.frame(r=rnorm(30))
ggplot(df)+
    stat_rose()

ggplot(df)+
    stat_rose(aes(petal=r,fill=r))+
    rose_text(aes(label=round(r)))

ggplot(df)+
    stat_rose(aes(petal=r,fill=r))+
    rose_text(aes(petal=r,label=round(r)))
}
