% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_rose.R
\docType{data}
\name{sfROse}
\alias{sfROse}
\alias{sf_rose}
\alias{StatSfRose}
\title{Simple feature style Rose Map}
\format{An object of class \code{StatSf} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
sf_rose(mapping = aes(petal = NULL), data = NULL, translate = c(0,
  0), rotate = 0, mirror = NULL, mirror_h = NULL, mirror_v = NULL,
  zoom = 1, position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)

StatSfRose
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{translate}{translate, must be two numbers. The first number is given to move along x-axis, the other is y-axis}

\item{rotate}{rotate}

\item{mirror}{mirror, must be two numbers. The first number is slope, the other is intercept}

\item{mirror_h}{one number, mirror refelct by horizontal line}

\item{mirror_v}{one number, mirror refelct by vertical line}

\item{zoom}{one number}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{passed to params}
}
\value{
Simple feature picture
}
\description{
Simple feature style Rose Map
}
\examples{

librar(ggrose)
dd=nCov2019()
dd2=rose_data(dd)

# original data
ggplot(data = dd2)+
    sf_rose()


# add petal
ggplot(data = dd2)+
    sf_rose(aes(petal=diagnose))

# add petal
# transform diagnose to log10
ggplot(data = dd2)+
    sf_rose(aes(petal=log10(diagnose)))


# add color
ggplot(data = dd2)+
    sf_rose(aes(fill=log10(diagnose)))

# use color and petal together
ggplot(data = dd2)+
    sf_rose(aes(fill=log10(diagnose),
                petal=log10(diagnose)))

# plot two roses and translate the second 1.6 anlong x-axis
ggplot(data = dd2)+
    sf_rose(aes(petal=log10(diagnose)))+
    sf_rose(aes(petal=log10(diagnose)),
            translate = c(1.6,0))

# rotate
ggplot(data = dd2)+
    sf_rose(aes(fill=log10(diagnose),
                petal=log10(diagnose)),
            rotate = 90)


# mirror reflecte
ggplot(data = dd2)+
    sf_rose(aes(petal=log10(diagnose)))+
    sf_rose(aes(petal=log10(diagnose)),
            translate = c(1.6,0),fill='blue',
            mirror_v = 2)

# zoom
ggplot(data = dd2)+
    sf_rose(aes(petal=log10(diagnose)))+
    sf_rose(aes(petal=log10(diagnose)),
            translate = c(1.6,0),fill='blue',
            zoom=0.5
            )

# mirror reflecte and regulate start angle
ggplot(data = dd2)+
    sf_rose(aes(petal=log10(diagnose)),rotate = -30)+
    sf_rose(aes(petal=log10(diagnose)),rotate = -30,
            translate = c(2,0),fill='blue',
            mirror_v = 2)
}
\keyword{datasets}
