% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_circle.R
\name{stat_circle}
\alias{stat_circle}
\title{Plot Circle}
\usage{
stat_circle(mapping = aes(), data = NULL, center = c(0, 0),
  r = 0.5, n = 1000, round = 360, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{center}{center coordinate, must be two numbers, default is 0,0}

\item{r}{radius, default is 1}

\item{n}{the number of points, which will build a circle to help locate the text}

\item{round}{the round of map, which should be 0-360, default is 360.}

\item{na.rm}{na rm, defalut is FALSE}

\item{show.legend}{show legend}

\item{inherit.aes}{inherit aes}

\item{...}{passed parameters}
}
\value{
ggplot2 style picture
}
\description{
Plot Circle
}
\examples{
library(ggrose)
set.seed(2020)
df=data.frame(r=rnorm(30))
ggplot(df)+
    stat_rose()

ggplot(df)+
    stat_rose(color='black',fill=NA)


ggplot(df)+
    stat_rose(aes(petal=r,fill=r))
}
