% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_rose.R
\name{stat_rose}
\alias{stat_rose}
\title{Rose Plot}
\usage{
stat_rose(mapping = aes(), data = NULL, center = c(0, 0),
  translate.1 = NULL, trans_h.2 = NULL, trans_v.3 = NULL,
  trans_al.4 = NULL, rotate.5 = 0, zoom.6 = NULL, mirror.7 = NULL,
  mirror_h.8 = NULL, mirror_v.9 = NULL, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, n = 1000, round = 360,
  outer.r = 1, inner.r = 0, inner.petal = FALSE, start_angle = 0,
  ...)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data, must be a dataframe}

\item{center}{center coordinate, must be two numbers}

\item{translate.1}{Operation to whole text, which is should be two numbers for x and y values}

\item{trans_h.2}{Operation to whole text, which is should be one number}

\item{trans_v.3}{Operation to whole text, which is should be one number}

\item{trans_al.4}{Operation to whole text, which is should be two numbers for slope and intercept values}

\item{rotate.5}{Operation to whole text, which is should be one number for angle}

\item{zoom.6}{Operation to whole text, which is should be one number for ratio}

\item{mirror.7}{Operation to whole text, which is should be two numbers for slope and intercept values}

\item{mirror_h.8}{Operation to whole text, which is should be one number}

\item{mirror_v.9}{Operation to whole text, which is should be one number}

\item{na.rm}{default is FALSE}

\item{show.legend}{default is FALSE}

\item{inherit.aes}{default is TRUE}

\item{n}{the number of points, which will build a circle to help locate the text}

\item{round}{the round of map, which should be 0-360, default is 360.}

\item{outer.r}{the radius of outer circle, default is 1}

\item{inner.r}{the radius of inner circle, default is 0}

\item{inner.petal}{a logical object. Defalut is FLASE. Whether do petal on inner circle.}

\item{start_angle}{start angle for rose, defalut is 0}

\item{...}{passed to parameters}
}
\value{
ggplot2 style picture
}
\description{
Rose Plot
}
\examples{
library(ggrose)
set.seed(2020)
df=data.frame(r=rnorm(30))


ggplot(df)+
    stat_circle(color='black',fill=NA)+
    stat_circle(r=1,color='black',fill=NA)+
    stat_circle(r=2,color='black',fill=NA)
}
