% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_rose.R
\docType{data}
\name{stat-rose}
\alias{stat-rose}
\alias{stat_rose}
\alias{StatRose}
\title{Rose Plot}
\format{An object of class \code{StatRose} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
stat_rose(mapping = aes(), data = NULL, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)

StatRose
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{na.rm}{na rm, defalut is FALSE}

\item{show.legend}{show legend}

\item{inherit.aes}{inherit aes}

\item{...}{passed parameters}
}
\value{
ggplot2 style picture
}
\description{
Rose Plot
}
\examples{
set.seed(2020)
df=data.frame(r=rnorm(30))
ggplot(df)+
    stat_rose()

ggplot(df)+
    stat_rose(color='black',fill=NA)


ggplot(df)+
    stat_rose(aes(petal=r,fill=r))
}
\keyword{datasets}
